% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acTFs.R
\name{GETacTFS}
\alias{GETacTFS}
\title{activated Transcription Factors}
\usage{
GETacTFS(DEGup, DEGdown, totalgenenum, FDRcutoff)
}
\arguments{
\item{DEGup}{The up-regulated DEG names(official symbol) provided as a list}

\item{DEGdown}{The down-regulated DEG names(official symbol) provided as a list}

\item{totalgenenum}{The total gene numbers used in DEG analysis}

\item{FDRcutoff}{the cutoff of FDR to define activated TFs}
}
\value{
allTFs: all the TFs and pvalues

sigTFS: significanly activated TFs and pvalues

sigTFsDEGnet: relations between the significanly activated TFs and their target DEGs (activation or repression).This relation list can be used for Cytoscape for visualization
}
\description{
This package is designed to identify significantly activated transcription factors (TFs) in the interested samples compared to the control group. Instead of comparing the expression of the TFs themselves, the significance of a TF activation is measured by Fisher’s exact test on its target genes. For a TF, both the upregulation of its activated genes and the downregulation of its repressed genes are considered as the enhancement of its function. First, users should perform DEG analysis (DEseq2 is recommended) between two groups of samples, then provide the total number of genes in the analysis and the name of the DEGs. The package includes a total of 3144 TF-target activation relationships and 1922 TF-target repression relationships derived from published articles for reference (Han H et al. Nucleic Acids Res. 2018;46(D1):D380-d6.). The "GETacTFS" function would exam the target genes changing of all 795 known TFs considering the regulation mode. The result includes the pvalue of each TF and a network of the significantly activated TFs and their targeted DEGs. To get the inactivated TFs in the treated group compared with the control group, you can swap the contents of DEGup and DEGdown and run GETacTFS again.
}
\examples{
# to get the significanly activated TFs in the treated group compared with the control group:
DEGup=read.table("up.txt",sep="\t",header=T,check.names=F)
DEGup=DEGup[,1]
DEGdown=read.table("down.txt",sep="\t",header=T,check.names=F)
DEGdown=DEGdown[,1]
result=GETacTFS(DEGup=DEGup,DEGdown=DEGdown,totalgenenum=20250,FDRcutoff=0.05)
head(result$allTFs)
head(result$sigTFS)
head(result$sigTFsDEGnet)
# To get the inactivated TFs in the treated group, just swap the contents of DEGup and DEGdown and run GETacTFS again.
DEGup=read.table("up.txt",sep="\t",header=T,check.names=F)
DEGup=DEGup[,1]
DEGdown=read.table("down.txt",sep="\t",header=T,check.names=F)
DEGdown=DEGdown[,1]
result=GETacTFS(DEGup=DEGdown,DEGdown=DEGup,totalgenenum=20250,FDRcutoff=0.05)
}
\author{
liguangqimed
}
