% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acTFs.R
\name{GETacTFS}
\alias{GETacTFS}
\title{activated Transcription Factors}
\usage{
GETacTFS(DEGup, DEGdown, totalgenenum, pcutoff)
}
\arguments{
\item{DEGup}{The up-regulated DEG names(official symbol) provided as a list}

\item{DEGdown}{The down-regulated DEG names(official symbol) provided as a list}

\item{totalgenenum}{The total gene numbers used in DEG analysis}

\item{pcutoff}{the cutoff of pvalue to define activated TFs}
}
\value{
allTFs: all the TFs and pvalues

sigTFS: significanly activated TFs and pvalues

sigTFsDEGnet: relations between the significanly activated TFs and their target DEGs (activation or repression).This relation list can be used for Cytoscape for visualization
}
\description{
activated Transcription Factors
}
\examples{
DEGup=c("EFCAB6","MAP2K6","GNPDA2")
DEGdown=c("LURAP1L","NPAP1","FGFR1")
result=GETacTFS(DEGup,DEGdown,totalgenenum=20000,pcutoff=0.05)
head(result$allTFs)
head(result$sigTFS)
head(result$sigTFsDEGnet)

DEGup=read.table("up.txt",sep="\t",header=T,check.names=F)
DEGup=DEGup[,1]
DEGdown=read.table("down.txt",sep="\t",header=T,check.names=F)
DEGdown=DEGdown[,1]
result=GETacTFS(DEGup,DEGdown,totalgenenum=20000,pcutoff=0.05)
head(result$allTFs)
head(result$sigTFS)
head(result$sigTFsDEGnet)
}
\author{
liguangqimed
}
